test commit

+++ tasks/example_processes/common/include/common.hpp
@@ -9,7 +9,7 @@ namespace nesterov_a_test_task_processes {
 
 using InType = int;
 using OutType = int;
-using TestType = std::tuple<int, std::string>;
+using TestType = std::tuple<double, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace nesterov_a_test_task_processes
